requirements.txt
requests==2.31.0
pytest==7.4.4
jsonschema==4.21.1
curlify==2.2.1
allure-pytest==2.13.2
python-dotenv==1.0.1